Refactored App.tsx and moved IModelButtonComponent, RenderViewportComponent to their own files. OpenIModelButton has been renamed to IModelButtonComponent and IModelComponents have been renamed to RenderViewportComponent.
